Improved packaging - only ZIPs of actually updated languages are replaced
As discussed with Eloy, it is responsibility of the packages producer to
decide what ZIP are updated. MD5 is always different for newly generated
ZIPs so the component's timemodified stamp is used as indicator.
Note that the timestamp of the newly generated package can be actually
lower than the previous one, if the most recently changed string was
removed from it.