Do not display 'mark as up-to-date' checkbox if the outdated string has not been committed yet
If the translated string is not committed and it becomes outdated (eg it
is stashed for some time, the English original is modified and then the
stash is applied back to the stage), we can not mark is as up-to-dat
because it is not in the repository yet. Instead, display just the
information that the string is outdated.