AMOS commit timestamp set to time() rather than to the Git commit's timestamp when parsing Moodle sources
Now when Moodle uses Git, commits can land with a timestamp in the past.
In AMOS, the commit's timestamp is used to determine the most recent
status of the string. Therefore, we have to use the time when the Git commit
was actually processed by AMOS rather than the original Git commit
timestamp (given that Git commits are processed in topological order).

The previous behaviour when AMOS commit's timestamp was set to the
original Git commit's timestamp worked well for a mirror of CVS where
the chronological and topological order was the same.