Dealing with commits having the same timestamp
Since recent changes, Git commits are processed in reverse topological
order and are recorded in AMOS repository with the current timestamp and
not the original Git's one. The hack in cli/parse-core.php to make sure
that two commits do not have the same timemodified stamp was buggy and
did not work. Because it was ugly anyway, I decided to stop relying on
such stupid requirement and do things properly.

This commit introduces two unit tests of from_snapshot() method to check
how it behaves when there are two repository records of the same string
with the same timemodified. The method's behaviour has been correct,
returning the more recently committed version - that is the one with
higher id.

The same logic was applied to two other places. AMOS translator's
constructor and timeline history viewer were fixed so they take the
record id into account, too.

After these fixes, the forced delay between commits during parse-core
processing is not needed any more again and was therefore dropped.