Dealing with commits having the same timestamp - part II
Yet another issue emerged - if there are two commits with the same
timemodified, one changes the string and deletes it. The commit
with higher id wins in this case, too. A unit test id added to make
sure that this logic has worked in from_snapshot(). Then, AMOS
translator is fixed so it respects this logic, too.

The condition deleted=0 had to go away from SQL and we deal with this
flag when generating the $s tree.