Fixed typo in reverse cleanup code
Due to this typo, the timemodified stamp of the removed string was set
to the timemodified value of the most recent version of the string.
Instead, the current time should be used. This did not influence the
functionality because the string removal record took precedence over the
previous record. However, it makes things more complicated when viewing
a history or re-playing it into Git.