AMOS repository checker
This CLI script is supposed to check data consistency and validity of
the AMOS repository. It is designed to be executed by the crontab, having the
standard output redirected into a log file and the standard error output
being sent automatically as a warning for the AMOS manager.

The first check implemented is the decsep/thousandssep check (MDL-31332)