Prevent empty components to be committed
This fixes a regression of d94c8da3. If in the block above all strings
from a component are unlinked, the empty component remains and the
macro is creating an empty commit. So we rebase the stage once more to
make sure that such empty components are dropped and has_component()
checks work.