Prevent commit loops due to reverted commits
The parse-core script now checks if the given git commit was already
processed and does not try to record the string changes again. This
helps to deal with situations where a string is changed and deleted
during one invocation of the script (which happens with reverted commits
for example). Without this check, such a string would be added, removed;
added, removed; added, removed; ... in each invocation.