Improve the tools for merging en_fix
Strings with the apostrophe were ignored by the previous version of the
script. There are still some cases (like double quoted strings with
apostrophe inside etc) that are not fully supported. But those are
pretty rare and can be fixed manually by observing the [WRN] messages in
the process log.