Automatically accept contributions if they have nothing to add
The reason behind this is that maintainers tend to manually accept the
contribution (which I can understand). So we just do it automatically if there
is a high chance that the contribution was just committed.

Note that there are certain situations where a contribution can be marked as
accepted by accident but that should be a rare case such as multiple
contributions all providing the same strings etc.