MDLSITE-3308 Create lang-table.html for download.moodle.org
This is a html table + generation timestamp which will be used in the
new generation of the downloads pages (MDLSITE-2784).

Note that I intentionally duplicated the code here and did not refactor
the existing renderer. We should eventually get rid of the php renderer,
but for the time being this is just a migration step to have both files
generated.