MDLSITE-5677 Fix the typing delay
There was ~250ms delay after each key pressed which made the actual
typing extremely uncomfortable.

The delay was caused by the node-event-simulate YUI3 module which
apparently traversed through the whole components table consisting of
40.000 cells, giving each of them an option to react on the event (which
none of them did).

The way to prevent the node-event-simulate module from being loaded on
this page, was to stop using the action_menu widget which was used in
the user menu element (top right corner). The actual hiding happens via
the theme, we just need to tell it to use the righjt layout for this
page.

Technically, we needed to hide just the user menu. But it seems to look
better if that corner is just empty.