MDLSITE-6243 Make mlang_versions computable
Previously we needed to change the code to add new versions. The idea is
to switch over to a few CFG settings (such as the list of version codes)
and the class should be able to deduct all the other properties, without
the need to touch the code to register new branches.