MDLSITE-6338 Fix evaluation of outdated strings
If a translation applies since a later version, consider it up to date.
These strings are technically still outdated but there is no trivial way
to fix this case of outdated translations that apply since more recent
versions.

This affects few strings only and it is a heritage of the previous AMOS
storage per-branch model. It should not cause any trouble for new
strings.