MDLSITE-6337 Improve export-zip processing to avoid memory problems
Frequent serialization and unserialization of the packinfo file
contributed a lot to the memory consumption. It turns out that we no
longer need it actually now as we have the stats manager.

Additionally, the plan is to run ZIP packs generation in smaller chunks,
allowing the OS to better cope with eventual PHP memory leaks. So now it
is the caller's responsibility to set the lastexportzip config as
needed.