MDLSITE-6941 Change behaviour of the MOV command
When the Git commit script contained the MOV command but the English
original was not actually moved from the file, it turned the string into
a non-editable state.

This change makes it so that the MOC command is not technically an alias
of the CPY. The language packs are interesect'ed against the English
original, so it should not be a problem to keep those strings.