MDLSITE-8099 Improve out of memory error reporting
There are situations when AMOS cannot process all the registered strings
and the filter must be specified more (e.g. by some components only or
specify the string identifier). Up until now, when too many strings are
found, the script runs out of memory and the ugly "undefined" error pops
up. This change introduces a nicer way to communicate to the user that
there is not enough memory to process the current filter setting.